add "extend your key from gpg" doc
This page is going to be linked to from the expiry bot emails
that are sent out to people who's keys are going to expire in 3
days.

The link in the email will be titled something like:
"Automatically extend and maintain your GPG key using Fluidkeys"

We know the person is using the keyserver (since that's how we
heard about them), so we must assume they'll want to keep their
key in the keyserver up to date. Suggesting they make a cron task
for this felt like the best solution for now.

We've also discussed turning off key rotation, but haven't come
to any conclusions.