add docs/remove-crontab-lines-macos-mojave
When migrating a macOS users system from using cron, to launchd
we attempt to remove the leftover lines from their crontab. This
may fail, as Fluidkeys may not have permission to do this.

If this is the case, we alert the user that it failed to remove
the lines and link them to this url.