apt repo: narrow scope of cryptographic authority (avoid "apt-key add -")
For modern debian and debian-derived systems, you can specify which
OpenPGP certificates are specifically allowed for which repositories.

So currently, if someone with access to debian or ubuntu signing keys
(or copies of repositories signed by those keys) was able to write to
https://dlownload.fluidkeys.com/dekstop/apt/, they would be able to
inject installable packages on end-user systems, even though the
fluidkeys signing keys are still protected.

Additionally, piping the fluidkeys APT repo OpenPGP certificate into
"apt-key add -", indicates that the fluidkeys organization is
acceptable to certify *all* repos on the system.  You probably don't
want that responsibility.

This arrangement fine on debian stretch (apt 1.4.9), which is
currently "oldstable", but doesn't work on jessie (apt 1.0.9.8.4).  I
think that's a reasonable tradeoff in 2019.  I believe it was added in
apt 1.1, so it should work in ubuntu xenial and later, but i haven't
tested it on that platform.

See also: https://wiki.debian.org/DebianRepository/UseThirdParty